+++ base/__init__.py
@@ -1,2 +1,2 @@
-VERSION = [1, 12, 7]
+VERSION = [1, 12, 8]
 __VERSION__ = '.'.join(map(str, VERSION))